Extract the test utility classes onto an upper level.
