+++ server/src/test/java/io/spine/server/aggregate/AggregateMessageDeliveryShould.java
@@ -28,8 +28,8 @@ import io.spine.server.BoundedContext;
 import io.spine.server.aggregate.given.AggregateMessageDeliveryTestEnv.DeliveryProject;
 import io.spine.server.aggregate.given.AggregateMessageDeliveryTestEnv.SingleShardProjectRepository;
 import io.spine.server.aggregate.given.AggregateMessageDeliveryTestEnv.TripleShardProjectRepository;
-import io.spine.server.delivery.given.MessageDeliveryTestEnv.ParallelDispatcher;
-import io.spine.server.delivery.given.MessageDeliveryTestEnv.ThreadStats;
+import io.spine.server.delivery.given.ParallelDispatcher;
+import io.spine.server.delivery.given.ThreadStats;
 import io.spine.server.transport.memory.InMemoryTransportFactory;
 import io.spine.server.transport.memory.SynchronousInMemTransportFactory;
 import io.spine.test.aggregate.ProjectId;
