+++ server/src/test/java/io/spine/server/delivery/given/ThreadStats.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2018, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package io.spine.server.delivery.given;
+
+import com.google.common.collect.HashMultimap;
+import com.google.common.collect.Multimap;
+import com.google.common.collect.Multimaps;
+import com.google.protobuf.Message;
+
+/**
+ * The statistics of threads, in which the entity has been processed.
+ *
+ * <p>Required in order to verify the sharding configuration.
+ *
+ * @param <I> the type of entity identifiers
+ * @author Alex Tymchenko
+ */
+public class ThreadStats<I extends Message> {
+
+    private final Multimap<Long, I> threadToId =
+            Multimaps.synchronizedMultimap(HashMultimap.<Long, I>create());
+
+    public void recordCallingThread(I id) {
+        final long currentThreadId = Thread.currentThread()
+                                           .getId();
+        threadToId.put(currentThreadId, id);
+    }
+
+    public Multimap<Long, I> getThreadToId() {
+        return Multimaps.unmodifiableMultimap(threadToId);
+    }
+
+    public void clear() {
+        threadToId.clear();
+    }
+}
