+++ server/src/test/java/io/spine/server/procman/PmMessageDeliveryShould.java
@@ -25,8 +25,8 @@ import io.spine.core.Event;
 import io.spine.core.Rejection;
 import io.spine.grpc.StreamObservers;
 import io.spine.server.BoundedContext;
-import io.spine.server.delivery.given.MessageDeliveryTestEnv.ParallelDispatcher;
-import io.spine.server.delivery.given.MessageDeliveryTestEnv.ThreadStats;
+import io.spine.server.delivery.given.ParallelDispatcher;
+import io.spine.server.delivery.given.ThreadStats;
 import io.spine.server.procman.given.PmMessageDeliveryTestEnv.DeliveryPm;
 import io.spine.server.procman.given.PmMessageDeliveryTestEnv.QuadrupleShardPmRepository;
 import io.spine.server.procman.given.PmMessageDeliveryTestEnv.SingleShardPmRepository;
