+++ server/src/test/java/io/spine/server/projection/ProjectionEventDeliveryShould.java
@@ -23,8 +23,8 @@ import io.spine.core.Ack;
 import io.spine.core.Event;
 import io.spine.grpc.StreamObservers;
 import io.spine.server.BoundedContext;
-import io.spine.server.delivery.given.MessageDeliveryTestEnv.ParallelDispatcher;
-import io.spine.server.delivery.given.MessageDeliveryTestEnv.ThreadStats;
+import io.spine.server.delivery.given.ParallelDispatcher;
+import io.spine.server.delivery.given.ThreadStats;
 import io.spine.server.projection.given.ProjectionEventDeliveryTestEnv.DeliveryProjection;
 import io.spine.server.projection.given.ProjectionEventDeliveryTestEnv.SingleShardProjectRepository;
 import io.spine.server.projection.given.ProjectionEventDeliveryTestEnv.TripleShardProjectRepository;
